Normative: Store strings or objects in time zone slot, remove getter
Refactor so that ZonedDateTime's [[TimeZone]] internal slot can now store
either a string (builtin time zone) or an object (custom time zone).

The .timeZone getter is replaced with a .timeZoneId getter that always
returns a string, and a .getTimeZone() method that always returns a new
object if the slot stored a string.

The operations ToTemporalTimeZoneIdentifier and ToTemporalTimeZoneObject
convert a [[TimeZone]] slot value to one or the other.

See: #1808