Normative: Copy options object in Plain{Date,MonthDay,YearMonth}.{from,p.with}
Also following the precedent set in #2447, this preserves consistency with
analogous from/with operations in {Plain,Zoned}DateTime that need to read
the overflow option twice in order to deal with time and date units
separately.

(We don't necessarily have to include this commit if others don't want it.
I'm including it for completeness because that's what we did in Richard's
change in #2447.)