Normative: Avoid calling user code in no-op round operations
If the given rounding options specify a smallest unit of nanoseconds and a
rounding increment of 1, then the rounding operation is a no-op. In that
case, return a copy of the duration, without actually performing the
rounding operation. This applies to all round(), until(), and since()
methods.

In the case of Duration.p.round(), other conditions must be fulfilled as
well for the operation to be a no-op: no balancing must take place.

In the case of Instant.p.round() and PlainTime.p.round(), no user code is
called either way, so implementors are free to make this optimization
anyway.

This optimization was already the case in PlainDate.p.since/until() and
PlainYearMonth.p.since/until(), so this makes the other since/until()
methods consistent with those.