Normative: Separate zoned and plain operations in RoundDuration
This converts a ZonedDateTime relativeTo into a PlainDateTime relativeTo
only when necessary, and only once. Previously, it could be converted up
to a few separate times in each operation, such as
UnbalanceDurationRelative, RoundDuration, and BalanceDurationRelative.
Since the conversion is user-visible, we don't want to perform it when not
necessary or perform it more times than necessary.

Closes: #2247