Normative: Don't observably iterate array in built-in calendar's fields()
NOTE: This commit doesn't actually do anything in this current branch,
because CalendarFields is never called with a String _calendar_. This just
illustrates the change, which will only take effect when #2482 is merged.

When calling CalendarFields with a built-in calendar, previously we would
create a JS Array from the passed List, and pass it to %Temporal.Calendar.
prototype.fields%, which would iterate it observably. So, instead we call
CalendarDateFields when the calendar is a built-in calendar, which doesn't
do anything observable at all.

See: #2289