Normative: Fast-path dateUntil() when difference largest unit is days
Introduce an operation DifferenceDate, which has a fast-path through the
ISO calendar (via DaysUntil) in the case where we would otherwise call
calendar.dateUntil() with largestUnit === "day". Also return a blank
duration immediately if the two dates are the same day.

Note that this makes it impossible for the following BalanceDuration call
in DifferenceISODateTime to throw. The dateUntil() method will no longer
be called for largestUnits of "day" or less. So dateDifference.[[Days]]
can be at most Number.MAX_VALUE, but timeDifference cannot balance into
more than one day and make it overflow in the BalanceDuration call.